requirements.txt
yt-dlp>=2024.1.0
openai-whisper>=20231117
transformers>=4.35.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
srt>=3.5.3
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
ffmpeg-python>=0.2.0